+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.3.1
+
+Features:
+  - Added calculating muzzle energy to the chronograph
+
 ## 1.3.0
 
 Bug Fixes: